Replace setfileicon with fileicon
setfileicon is available as PPC and x86 binary but not for x86_64. Since
on macOS 10.15 Catalina only supports x86_64 architecture a switch to
another tool was necessary.

Use kuhmuh's commit 172ef51/"Replace setfileicon by fileicon". Beside
switching to fileicon kuhmuh enhanced the install script so that a
missing Firefox icon is automatically created (thanks for that!).

Update README.md to reflect the use of fileicon, it is available on
homebrew.

Removed static generated Firefox version pictures for Firefox version
60 and up. Also remove the setfileicon binary